MUUUUUCH better. Important specifics below: Ok, first of all I had to mess with the <link> or <script> in main.handlebars because it wouldn't render the local files.
I had to edit the path to our local CSS file in order to get it to work. I hope that doesn't screw things up for you guys on the server-side of things (but I figured it was unlikely so I went ahead with it).

Also I found it much much easier to just pull items off the web rather than local sources (for instance the Bootstrap.min.css file, as well as the cool background photo in the back).

I hope that the changes I made on members.handlebars will be pretty easily done to all the other pages.

NAVBAR: I started making some changes to the navbar, but I tried to keep any id's the same. please let me know if I screwed up something on the server-side due to any changes here.

Next steps:
1. getting the rest of the handlebars pages to look similar
2. getting that navbar to be snazzy looking but more importantly the items/links in it to be intuitive for the user.  I'm guessing there's potential for the different items on that nav menu being confuzing/users not knowing where they go exactly.
3. Getting everything to be mobile friendly.  I think by far the easiest way to get this is to really really lean on Bootstrap.  I'll probably need to take some of the CSS formatting I did manually and change it to be done through Bootstrap classes and stuff like that.
4. Whatever you guys can think of what I'm forgetting?...